Move client folder to post and rename Client to PostClient
