+++ src/post/client/types.ts
@@ -1,6 +1,6 @@
-import { Post, PostData } from "../post/types";
+import { Post, PostData } from "../types";
 
-export interface ClientStructure {
+export interface PostClientStructure {
   getPosts: () => Promise<{ posts: Post[] }>;
   createPost: (postData: PostData) => void;
 }
